+++ TheftOfTheWinterStar/docs/release-notes.md
@@ -2,8 +2,10 @@
 
 # Release notes
 ## Upcoming release
-* Fixed Tempus Globes not working on the Ginger Island farm.
-* Fixed Tempus Globes not protecting out-of-season crops on season change in some cases.
+* Fixed Tempus Globes...
+  * not working on the Ginger Island farm;
+  * not protecting crops on season change in some cases;
+  * still protecting crops in seasons for which they were previously covered by a Tempus Globe even if it was later removed.
 
 ## 1.0.2
 Released 19 June 2021 for Stardew Valley 1.5. Updated by Pathoschild.